project restructure continued
