+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import serverPkg "github.com/21Philip/Auction/server"
+import serverPkg "github.com/21Philip/Auction/internal/server"
 
 func main() {
 	serverPkg.StartServer(5)
